24.05.05 BOJ G4 Graph 1 fail

+++ Adhoc/S5_31738.java
@@ -23,7 +23,7 @@ public class S5_31738 {
             return;
         }
 
-        int res = 1;
+        long res = 1;
         for (int i = 1; i <= N; i++) {
             res = (res * i) % M;
         }
